remove the interface
